third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/namespaced-elements-and-selectors.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id="sandbox">
    <span></span>
</div>
<script>
description("querySelectorAll and namespaces");

sandbox.appendChild(document.createElementNS("http://dummynamespace", "span"));
sandbox.appendChild(document.createElementNS("", "span"));

shouldBe("sandbox.querySelectorAll('span').length", "3");
shouldBe("sandbox.querySelectorAll('*|span').length", "3");
shouldBe("sandbox.querySelectorAll('|span').length", "1");
</script>
