third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/namespaced-elements-and-selectors-expected.txt

querySelectorAll and namespaces

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sandbox.querySelectorAll('span').length is 3
PASS sandbox.querySelectorAll('*|span').length is 3
PASS sandbox.querySelectorAll('|span').length is 1
PASS successfullyParsed is true

TEST COMPLETE
